Fix push/pop/call/ret to always keep stack layout
We now capture rsp changes as required, as we
cannot change the stack layout in DBrew; this is
out-of-scope for binary transformation.

Add test case for correct push/pop stack handling,
and adjust expected test output accordingly.

This also fixes test cases/integration/op-push-pop.s to
make it deterministic: "pop bx" only sets to lower 16 bits,
leaving the others untouched, but rbx was not initialized
to a fixed value before.

TODO: With a lot of push/pop, it now becomes obvious that
we want to do some cleanup passes (merge rsp changes, remove
dead code, ...)